Context menu 5: refactor into sub-files
